requirements.txt
assemblyai>=0.20.0
yt-dlp>=2024.1.0
sentence-transformers>=2.7.0
numpy>=1.26.0
scipy>=1.12.0
jinja2>=3.1.0
python-dotenv>=1.0.0
librosa>=0.10.0
pyannoteai-sdk>=0.4.0
